move dashboard type into seperate files
